Adding Admin Controller

+++ src/main/java/com/vividhdarote/journalismApp/service/JournalEntryService.java
@@ -21,38 +21,49 @@ public class JournalEntryService {
     @Autowired
     private UserService userService;
 
-    public List<JournalEntry> getAll(){
+    public List<JournalEntry> getAll() {
         return journalEntryRepository.findAll();
     }
 
     @Transactional
-    public void saveEntry(JournalEntry journalEntry, String userName){
-        try{
-            User user= userService.findByUserName(userName);
+    public void saveEntry(JournalEntry journalEntry, String userName) {
+        try {
+            User user = userService.findByUserName(userName);
             journalEntry.setDate(LocalDateTime.now());
             JournalEntry saved = journalEntryRepository.save(journalEntry);
             user.getJournalEntryList().add(saved);
-            userService.saveEntry(user);
-        }
-        catch (Exception e){
+            userService.saveUser(user);
+        } catch (Exception e) {
             System.out.println(e);
-            throw new RuntimeException("An error occurred while saving the data: "+e);
+            throw new RuntimeException("An error occurred while saving the data: " + e);
         }
     }
 
-    public void saveEntry(JournalEntry journalEntry){
+    public void saveEntry(JournalEntry journalEntry) {
         journalEntryRepository.save(journalEntry);
     }
 
-    public Optional<JournalEntry> findById(ObjectId id){
+    public Optional<JournalEntry> findById(ObjectId id) {
         return journalEntryRepository.findById(id);
     }
 
-    public void deleteById(ObjectId id, String userName){
-        User user= userService.findByUserName(userName);
-        user.getJournalEntryList().removeIf(x-> x.getId().equals(id));
-        userService.saveEntry(user);
-        journalEntryRepository.deleteById(id);
+    @Transactional
+    public boolean deleteById(ObjectId id, String userName) {
+        boolean removed = false;
+        try {
+            User user = userService.findByUserName(userName);
+            removed = user.getJournalEntryList().removeIf(x -> x.getId().equals(id));
+            if (removed) {
+                userService.saveUser(user);
+                journalEntryRepository.deleteById(id);
+            }
+        } catch (Exception e) {
+            System.out.println(e);
+            throw new RuntimeException("An error occurred while deleting the entry: ", e);
+        }
+        return removed;
+
+
     }
 
 }
